feat(@embark/specialconfigs): introduce dynamic address support
This commit introduces a new Smart Contract configuration `addressHandler`
that lets users define a function to "lazily" compute the address of the
Smart Contract in question.

This is useful when a third-party takes care of deploying a dependency
Smart Contract, but the address of that Smart Contract only being available
at run-time.

Example:

```
contracts: {
  SimpleStorage: {
    fromIndex: 0,
    args: [100],
  },
  OtherContract: {
    deps: ['SimpleStorage'],
    addressHandler: async (deps) => {
      // use `deps.contracts.SimpleStorage` to determine address
    },
    abiDefinition: ABI
  },
}
```

In the example above, `OtherContract` will be deployed after `SimpleStorage`
because it uses the `deps` property to define Smart Contracts that it depends
on. All dependencies are exposed on the `addressHandler` function parameter
similar to deployment hooks.

**Note**: This only works when an ABI is provided for the Smart Contract that
is already deployed (in this case `OtherContract`).

Closes #1690